requirements.txt
discord.py[voice]
lark
